+++ tests/scalers_go/rabbitmq/rabbitmq_queue_http/rabbitmq_queue_http_test.go
@@ -13,6 +13,7 @@ import (
 	"k8s.io/client-go/kubernetes"
 
 	. "github.com/kedacore/keda/v2/tests/helper"
+	. "github.com/kedacore/keda/v2/tests/scalers_go/rabbitmq"
 )
 
 // Load environment variables from .env file
